# scripts/contracts/v1/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeChangeRequest": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Author",
            "type": "string"
          },
          "branch": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "changed_services": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Changed Services",
            "type": "array"
          },
          "commit_sha": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 7,
                "pattern": "^[0-9a-fA-F]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "deployment_environment": {
            "default": "staging",
            "maxLength": 80,
            "minLength": 1,
            "title": "Deployment Environment",
            "type": "string"
          },
          "files_changed": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Files Changed",
            "type": "integer"
          },
          "flags": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 100,
            "title": "Flags",
            "type": "array"
          },
          "lines_added": {
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Lines Added",
            "type": "integer"
          },
          "lines_deleted": {
            "maximum": 10000000.0,
            "minimum": 0.0,
            "title": "Lines Deleted",
            "type": "integer"
          },
          "observability_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Observability Score",
            "type": "number"
          },
          "previous_failures": {
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Previous Failures",
            "type": "integer"
          },
          "repository": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Repository",
            "type": "string"
          },
          "rollback_ready": {
            "title": "Rollback Ready",
            "type": "boolean"
          },
          "test_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Test Coverage",
            "type": "number"
          },
          "title": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "repository",
          "author",
          "files_changed",
          "lines_added",
          "lines_deleted",
          "changed_services",
          "test_coverage",
          "rollback_ready",
          "observability_score",
          "previous_failures"
        ],
        "title": "AnalyzeChangeRequest",
        "type": "object"
      },
...
      "ChangeDetail": {
        "additionalProperties": false,
        "properties": {
          "analysis_schema_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Analysis Schema Version",
            "type": "string"
          },
          "author": {
            "title": "Author",
            "type": "string"
          },
          "blast_radius": {
            "$ref": "#/components/schemas/BlastRadius"
          },
          "branch": {
            "title": "Branch",
            "type": "string"
          },
          "changed_services": {
            "items": {
              "type": "string"
            },
            "title": "Changed Services",
            "type": "array"
          },
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_mode": {
            "enum": [
              "synthetic",
              "connected"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "deployment_environment": {
            "title": "Deployment Environment",
            "type": "string"
          },
          "deployment_status": {
            "title": "Deployment Status",
            "type": "string"
          },
          "engine_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "files_changed": {
            "minimum": 0.0,
            "title": "Files Changed",
            "type": "integer"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "graph_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Graph Version",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lines_added": {
            "minimum": 0.0,
            "title": "Lines Added",
            "type": "integer"
          },
          "lines_deleted": {
            "minimum": 0.0,
            "title": "Lines Deleted",
            "type": "integer"
          },
          "repository": {
            "title": "Repository",
            "type": "string"
          },
          "repository_id": {
            "title": "Repository Id",
            "type": "string"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskLedger"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "scoring_policy_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Scoring Policy Version",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "workspace_id": {
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspace_id",
          "repository_id",
          "scenario_id",
          "data_mode",
          "analysis_schema_version",
          "engine_version",
          "scoring_policy_version",
          "graph_version",
          "title",
          "repository",
          "author",
          "commit_sha",
          "branch",
          "created_at",
          "deployment_status",
          "deployment_environment",
          "changed_services",
          "files_changed",
          "lines_added",
          "lines_deleted",
          "flags",
          "risk",
          "blast_radius"
        ],
        "title": "ChangeDetail",
        "type": "object"
      },
...
      "FeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "hypothesis_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Hypothesis Id",
            "type": "string"
          },
          "note": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Note",
            "type": "string"
          },
          "verdict": {
            "enum": [
              "confirmed",
              "rejected",
              "partial"
            ],
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "hypothesis_id",
          "verdict",
          "note"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentDetail": {
        "additionalProperties": false,
        "properties": {
          "affected_services": {
            "items": {
              "type": "string"
            },
            "title": "Affected Services",
            "type": "array"
          },
          "analysis_schema_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Analysis Schema Version",
            "type": "string"
          },
          "assignee_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          },
          "correlated_change_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlated Change Id"
          },
          "data_mode": {
            "enum": [
              "synthetic",
              "connected"
            ],
            "title": "Data Mode",
            "type": "string"
          },
          "engine_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Engine Version",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/Feedback"
            },
            "title": "Feedback",
            "type": "array"
          },
          "graph_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Graph Version",
            "type": "string"
          },
          "hypotheses": {
            "items": {
              "$ref": "#/components/schemas/Hypothesis"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "scoring_policy_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Scoring Policy Version",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "title": "Timeline",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "data_mode",
          "analysis_schema_version",
          "engine_version",
          "scoring_policy_version",
          "graph_version",
          "title",
          "severity",
          "status",
          "started_at",
          "resolved_at",
          "affected_services",
          "correlated_change_id",
          "summary",
          "timeline",
          "evidence",
          "hypotheses",
          "feedback"
        ],
        "title": "IncidentDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DeployGuard AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/changes": {
      "get": {
        "operationId": "changes_api_v1_changes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChangeDetail"
                  },
                  "title": "Response Changes Api V1 Changes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Changes"
      }
    },
    "/api/v1/changes/analyze": {
      "post": {
        "operationId": "change_analyze_api_v1_changes_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Analyze"
      }
    },
    "/api/v1/changes/{change_id}": {
      "get": {
        "operationId": "change_api_v1_changes__change_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "change_id",
            "required": true,
            "schema": {
              "title": "Change Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change"
      }
    },
...
    "/api/v1/incidents": {
      "get": {
        "operationId": "incidents_api_v1_incidents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentDetail"
                  },
                  "title": "Response Incidents Api V1 Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incidents"
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incident"
      }
    },
    "/api/v1/incidents/{incident_id}/export-markdown": {
      "get": {
        "operationId": "incident_export_markdown_api_v1_incidents__incident_id__export_markdown_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incident Export Markdown"
      }
    },
    "/api/v1/incidents/{incident_id}/feedback": {
      "post": {
        "operationId": "incident_feedback_api_v1_incidents__incident_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incident Feedback"
      }
    },
...
    "/api/v1/incidents/{incident_id}/synthesize-llm": {
      "post": {
        "operationId": "incident_synthesize_llm_api_v1_incidents__incident_id__synthesize_llm_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LLMSynthesisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Incident Synthesize Llm"
      }
    },
...
    "/api/v1/overview": {
      "get": {
        "operationId": "overview_api_v1_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview"
      }
    },
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "scenarios_api_v1_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioSummary"
                  },
                  "title": "Response Scenarios Api V1 Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scenarios"
      }
    },
    "/api/v1/scenarios/{scenario_id}/activate": {
      "post": {
        "operationId": "scenario_activate_api_v1_scenarios__scenario_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Scenario Activate"
      }
    },
...
    "/api/v1/telemetry/events": {
      "post": {
        "operationId": "telemetry_ingest_api_v1_telemetry_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelemetryIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Telemetry Ingest Api V1 Telemetry Events Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Telemetry Ingest"
      }
    },
...
  }
}